Use bookmark title throughout UI
