+++ src/components/series-row.js
@@ -34,7 +34,7 @@ const SeriesRow = ({ feedItem: item, onMoreClick, ...props }: Props) => {
         />
         <div className="of-hidden">
           <div className="fs-16 fs-20-m fw-semibold lh-1d25 of-hidden to-ellipsis ws-noWrap">
-            {item.series.title}
+            {item.title}
           </div>
           <div className="fs-12 fs-14-m o-50p">{item.series.site.name}</div>
         </div>
